Reorganise some code from routes into controllers/handler
