+++ application/.eslintrc
@@ -34,6 +34,7 @@
     "no-unexpected-multiline": 2,
     "no-var": 1,
     "no-warning-comments": 0,
+    "no-undef": 2,
     "promise/always-return": "warn",
     "promise/no-return-wrap": "error",
     "promise/param-names": "error",
